SW-2814 Move search code to services directory - no functional change here - after this change, we write other services that use the underlying search service  (eg.) seedbank service -> getSeedbanks() [uses search beneath]
