+++ src/services/index.ts
@@ -3,6 +3,7 @@ import HttpService from './HttpService';
 import OrganizationUserService from './OrganizationUserService';
 import OrganizationService from './OrganizationService';
 import PreferencesService from './PreferencesService';
+import SearchService from './SearchService';
 import SystemService from './SystemService';
 import UserService from './UserService';
 
@@ -14,6 +15,7 @@ export {
   OrganizationService,
   OrganizationUserService,
   PreferencesService,
+  SearchService,
   SystemService,
   UserService,
 };
